+++ TVING_Clone/TVING_Clone/Global/Base/BaseUIView.swift
@@ -1,0 +1,32 @@
+//
+//  BaseUIView.swift
+//  TVING_Clone
+//
+//  Created by 박윤빈 on 2023/05/04.
+//
+
+import UIKit
+
+class BaseUIView: UIView {
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configureUI()
+        setLayout()
+    }
+    
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func configureUI() {
+        
+    }
+    
+    func setLayout() {
+        
+    }
+}
+
+
