+++ TVING_Clone/TVING_Clone/Global/Base/BaseViewController.swift
@@ -1,0 +1,59 @@
+//
+//  BaseViewController.swift
+//  TVING_Clone
+//
+//  Created by 박윤빈 on 2023/05/04.
+//
+
+import UIKit
+
+import SnapKit
+
+class BaseViewController: UIViewController {
+    
+    // MARK: - Properties
+    
+    lazy private(set) var className: String = {
+      return type(of: self).description().components(separatedBy: ".").last ?? ""
+    }()
+    
+    // MARK: - Initializing
+    
+    override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    deinit {
+        print("DEINIT: \(className)")
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        configureUI()
+        setLayout()
+        setButtonEvent()
+    }
+    
+    //MARK: - Functions
+    
+    func configureUI() {
+        //override Point
+        view.backgroundColor = .black
+
+    }
+    
+    func setLayout() {
+        //override Point
+    }
+    
+    func setButtonEvent() {
+        //override Point
+    }
+}
+
+
